requirements.txt
streamlit==1.28.0
msal==1.24.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.0
plotly==5.17.0
python-docx==0.8.11
PyPDF2==3.0.1
geopy==2.4.0
pymongo==4.5.0
numpy==1.24.3
faiss-cpu==1.7.4
scikit-learn==1.3.0
sentence-transformers==2.2.2